third_party/blink/web_tests/platform/generic/fullscreen/orthogonal-writing-mode-full-screen-crash-expected.txt

entering fullscreen OK
EVENT(webkitfullscreenchange)
entered fullscreen OK
EVENT(webkitfullscreenchange)
exited fullscreen OK
END OF TEST
